feat: add product schema and validator response error
